+++ src/pages/ChatRoom.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { ChatRoomPage } from '../components/Chat/ChatRoomPage';
+import { ChatRoomPage } from '../components/chating/ChatRoomPage';
 
 export const ChatRoom = () => {
   return (
