+++ src/pages/PartyRequestApprove.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { ChatApprove } from '../components/Chat/ChatApprove';
+import { ChatApprove } from '../components/chating/ChatApprove';
 
 export const PartyRequestApprove = () => {
   return <ChatApprove></ChatApprove>;
